apps/backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
sqlparse==0.5.3
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.0
redis==5.0.1
django-redis==5.4.0
drf-spectacular==0.27.0
python-decouple==3.8
